fix(view): deduplicate usage builder, constant and helpers (PHNX-3940)
Move viewUsageSummaryOptions into lib/accounting/usage.ts and have
view.ts + account-catalog.ts share the single builder (headless,
unverified, expectedWindows, errorKind, benignState, grok label) via
UsageInfo. Keep OVERVIEW_MAX_USAGE_WINDOWS in account-catalog.ts and
import it in view.ts. Import stripAnsi from lib/text/width.ts and drop
the local copy. Make usageText explicit: snapshot rows render through
formatUsageSummary only, non-snapshot rows through the legacy bar; remove
the fallback band-aid and trimStart hack. Replace inline
import('./accounting/usage.js').UsageSnapshot type expressions with proper
import type at the top of each file.

Fixes findings 1-5; tracks PHNX-3940.
